+++ packages/autolink/src/EmailMatcher.ts
@@ -4,12 +4,12 @@
  */
 
 import React from 'react';
-import { Matcher, MatchResponse } from 'interweave';
+import { Matcher, MatchResponse, Node } from 'interweave';
 import Email, { EmailProps } from './Email';
 import { EMAIL_PATTERN } from './constants';
 
 export default class EmailMatcher extends Matcher<EmailProps> {
-  replaceWith(match: string, props: EmailProps) {
+  replaceWith(match: string, props: EmailProps): Node {
     return React.createElement(Email, props, match);
   }