+++ packages/autolink/src/HashtagMatcher.ts
@@ -4,12 +4,12 @@
  */
 
 import React from 'react';
-import { Matcher, MatchResponse } from 'interweave';
+import { Matcher, MatchResponse, Node } from 'interweave';
 import Hashtag, { HashtagProps } from './Hashtag';
 import { HASHTAG_PATTERN } from './constants';
 
 export default class HashtagMatcher extends Matcher<HashtagProps> {
-  replaceWith(match: string, props: HashtagProps) {
+  replaceWith(match: string, props: HashtagProps): Node {
     return React.createElement(Hashtag, props, match);
   }